backend/requirements.txt
aiohttp
aiosignal
alembic
annotated-types
anyio
asgiref
async-timeout
attrs
backoff
bcrypt
beautifulsoup4
black
boto3
botocore
build
cachetools
certifi
cffi
charset-normalizer
chroma-hnswlib
chromadb
click
click-spinner
coloredlogs
croniter
cryptography
dataclasses-json
Deprecated
dirtyjson
distro
dnspython
email_validator
et-xmlfile
exceptiongroup
fastapi
fastapi-cli
fastapi-utilities
filelock
flatbuffers
frozenlist
fsspec
google-auth
googleapis-common-protos
greenlet
grpcio
h11
httpcore
httptools
httpx
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
iniconfig
Jinja2
jmespath
joblib
jwcrypto
kubernetes
llama-index
llama-index-agent-openai
llama-index-cli
llama-index-core
llama-index-experimental
llama-index-indices-managed-llama-cloud
llama-index-legacy
llama-index-llms-openai
llama-index-llms-gemini
llama-index-llms-anthropic
llama-index-multi-modal-llms-openai
llama-index-program-openai
llama-index-question-gen-openai
llama-index-readers-file
llama-index-readers-llama-parse
llama-index-storage-chat-store-redis
llama-index-vector-stores-chroma
llama-parse
llamaindex-py-client
loguru
Mako
markdown-it-py
MarkupSafe
marshmallow
mdurl
mmh3
monotonic
mpmath
multidict
mypy-extensions
nest-asyncio
networkx
nltk
numpy
oauthlib
onnxruntime
openai
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
overrides
packaging
pandas
pathspec
pillow
platformdirs
pluggy
posthog
protobuf
psycopg2-binary
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
Pygments
PyJWT
pypdf
PyPika
pyproject_hooks
pytest
python-dateutil
python-dotenv
python-multipart
pytz
PyYAML
redis
regex
requests
requests-oauthlib
rich
rsa
s3transfer
shellingham
six
sniffio
soupsieve
SQLAlchemy
starlette
striprtf
sympy
tenacity
tiktoken
tokenizers
tomli
tqdm
typer
typing-inspect
typing_extensions
tzdata
ujson
urllib3
uvicorn
watchfiles
websocket-client
websockets
wrapt
yarl
zipp
